Prevent editing another user's notes
* Raise 404 error if user tries to edit another user's note
* Filter notes list to those authored by current user
* Refactor email tip code into separate module
