+++ app/blueprints/notes/views.py
@@ -7,8 +7,7 @@ import datetime
 
 from flask import (
     Blueprint,
-    after_this_request,
-    current_app,
+    abort,
     jsonify,
     redirect,
     render_template,
@@ -17,7 +16,9 @@ from flask import (
 from flask_login import current_user
 from flask.ext.security import login_required
 from sqlalchemy import desc
+from sqlalchemy.orm.exc import MultipleResultsFound, NoResultFound
 
+from app.blueprints.notes.email_tip import EmailTip
 from app.blueprints.notes.models import Note
 
 
@@ -28,64 +29,32 @@ def two_mins_ago():
     return datetime.datetime.utcnow() - datetime.timedelta(minutes=2)
 
 
-class EmailTip(object):
-
-    def __init__(self):
-        current_app.logger.debug('EmailTip constructed')
-        self._times_seen = None
-        self.set_template_context()
-
-    @property
-    def times_seen(self):
-        if self._times_seen is None:
-            self._times_seen = int(request.cookies.get('seen_email_tip', 0))
-            current_app.logger.debug(
-                'reading cookie: {}'.format(self._times_seen))
-
-        return self._times_seen
-
-    @times_seen.setter
-    def times_seen(self, value):
-        current_app.logger.debug('set times_seen to {}'.format(value))
-        self._times_seen = value
-        after_this_request(self.set_cookie)
-
-    @property
-    def visible(self):
-        current_app.logger.debug('visible {}'.format(self.times_seen < 2))
-        return self.times_seen < 2
-
-    def incr_times_seen(self):
-        if self.visible:
-            current_app.logger.debug('incr_times_seen')
-            self.times_seen += 1
-
-    def set_cookie(self, response):
-        current_app.logger.debug('set_cookie: {}'.format(self.times_seen))
-        response.set_cookie('seen_email_tip', str(self.times_seen))
-        return response
-
-    def set_template_context(self):
-        current_app.jinja_env.globals['email_tip_visible'] = self.visible
-        current_app.logger.debug(
-            'updated template global email_tip_visible={}'.format(
-                self.visible))
-
-
 @notes.context_processor
 def notes_context():
     return {'undo_timeout': two_mins_ago}
 
 
+def get_or_404(class_, **kwargs):
+    try:
+        return class_.query.filter_by(**kwargs).one()
+
+    except NoResultFound:
+        abort(404)
+
+    except MultipleResultsFound:
+        raise
+
+
 @notes.route('/notes')
 @login_required
 def list():
-    all_notes = Note.query.order_by(desc(Note.updated)).all()
+    notes = Note.query.filter(Note.author == current_user)
+    notes = notes.order_by(desc(Note.updated)).all()
 
     email_tip = EmailTip()
     email_tip.incr_times_seen()
 
-    return render_template('notes/list.html', notes=all_notes)
+    return render_template('notes/list.html', notes=notes)
 
 
 @notes.route('/notes', methods=['POST'])
@@ -101,7 +70,6 @@ def add():
 
 @notes.route('/notes/dismiss-email-tip')
 def dismiss_tip():
-
     email_tip = EmailTip()
     email_tip.times_seen = 2
 
@@ -111,7 +79,7 @@ def dismiss_tip():
 @notes.route('/notes/<id>/undo', methods=['POST'])
 @login_required
 def undo(id):
-    note = Note.query.get(id)
+    note = get_or_404(Note, id=id, author=current_user)
 
     try:
         note.revert()
@@ -125,7 +93,7 @@ def undo(id):
 @notes.route('/notes/<id>/edit', methods=['POST'])
 @login_required
 def edit(id):
-    note = Note.query.get(id)
+    note = get_or_404(Note, id=id, author=current_user)
 
     note.update(request.form['content'])
